Undo refacto
